divide binding by module
